rm creator_id

+++ back-end/src/chat-gateway/chat.gateway.ts
@@ -77,10 +77,10 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
     }
   }
 
-  switchChannel(client: any, to?: string) {
+  async switchChannel(client: any, to?: string) {
     if (to != undefined)
-      this.activeService.updateUser({ id: client.id, chat_id: to }).then();
-    else this.activeService.updateUser({ id: client.id, chat_id: null }).then();
+      await this.activeService.updateUser({ id: client.id, chat_id: to }).then();
+    else await this.activeService.updateUser({ id: client.id, chat_id: null }).then();
     let tmp = this.rooms.get(client.id);
     if (tmp != '') {
       client.leave(tmp);
@@ -262,7 +262,7 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
     if (!user || !user.chat_id)
       return ;
     const users = await this.activeService.getUsersByChatId(user.chat_id);
-    this.switchChannel(client);
+    await this.switchChannel(client);
     if (users.length == 1)
       await this.chanService.deleteChannel(user.user_id, user.chat_id);
     else
@@ -270,12 +270,15 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
       const chan = await this.chanService.getChannelById(user.chat_id);
       if (!chan)
         return ;
-      this.chanService.deleteModerator(chan.creator_id, {chat_id: user.chat_id, user_id: user.user_id})
-      if (chan.creator_id)
+      const tmp = users.filter(val => val.id == user.user_id).length;
+      
+      if (tmp == 1)
       {
-        console.log("deleting creator on " + chan.id);
-        
-        await this.chanService.deleteCreator(chan.id)
+        await this.chanService.deleteModerator(chan.creator_id, {chat_id: user.chat_id, user_id: user.user_id})
+        if (chan.creator_id && user.user_id == chan.creator_id)
+        {             
+          await this.chanService.deleteCreator(chan.id)
+        }
       }
     }
